+++ src/lib/Scenes/Consignments/v2/Screens/ConsignmentsHome/Components/TextContainer.tsx
@@ -1,0 +1,6 @@
+import { Box } from "@artsy/palette"
+import styled from "styled-components/native"
+
+export const TextContainer = styled(Box)`
+  flex: 1;
+`
